+++ pulsar-testclient/src/main/java/org/apache/pulsar/testclient/PerformanceProducer.java
@@ -18,47 +18,11 @@
  */
 package org.apache.pulsar.testclient;
 
-import static java.util.Objects.requireNonNull;
-import static java.util.concurrent.TimeUnit.NANOSECONDS;
 import static org.apache.commons.lang3.StringUtils.isNotBlank;
-import static org.apache.pulsar.client.impl.conf.ProducerConfigurationData.DEFAULT_BATCHING_MAX_MESSAGES;
-import static org.apache.pulsar.client.impl.conf.ProducerConfigurationData.DEFAULT_MAX_PENDING_MESSAGES;
-import static org.apache.pulsar.client.impl.conf.ProducerConfigurationData.DEFAULT_MAX_PENDING_MESSAGES_ACROSS_PARTITIONS;
-import static org.apache.pulsar.testclient.PerfClientUtils.LATENCY_HISTOGRAM_SIGNIFICANT_DIGITS;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.ObjectWriter;
-import com.google.common.collect.Range;
-import com.google.common.util.concurrent.RateLimiter;
-import io.netty.util.concurrent.DefaultThreadFactory;
-import java.io.FileOutputStream;
-import java.io.PrintStream;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.time.Duration;
 import java.time.Instant;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-import java.util.Random;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.concurrent.Future;
-import java.util.concurrent.Semaphore;
-import java.util.concurrent.ThreadLocalRandom;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicLong;
-import java.util.concurrent.atomic.AtomicReference;
-import java.util.concurrent.atomic.LongAdder;
-import lombok.CustomLog;
-import org.HdrHistogram.Histogram;
-import org.HdrHistogram.HistogramLogWriter;
-import org.HdrHistogram.Recorder;
-import org.apache.pulsar.client.admin.PulsarAdmin;
-import org.apache.pulsar.client.admin.PulsarAdminBuilder;
-import org.apache.pulsar.client.admin.PulsarAdminException;
+import java.util.concurrent.CompletableFuture;
 import org.apache.pulsar.client.api.v5.Producer;
 import org.apache.pulsar.client.api.v5.ProducerBuilder;
 import org.apache.pulsar.client.api.v5.PulsarClient;
@@ -77,394 +41,44 @@ import org.apache.pulsar.client.api.v5.config.ProducerAccessMode;
 import org.apache.pulsar.client.api.v5.config.ProducerEncryptionPolicy;
 import org.apache.pulsar.client.api.v5.config.TransactionPolicy;
 import org.apache.pulsar.client.api.v5.schema.Schema;
-import org.apache.pulsar.common.partition.PartitionedTopicMetadata;
-import org.apache.pulsar.common.util.FutureUtil;
 import picocli.CommandLine.Command;
-import picocli.CommandLine.ITypeConverter;
 import picocli.CommandLine.Option;
-import picocli.CommandLine.TypeConversionException;
 
 /**
- * A client program to test pulsar producer performance.
+ * A client program to test pulsar producer performance with the V5 client API.
+ *
+ * <p>The V5 client is used so the command works transparently against both regular and scalable
+ * topics. Everything that is not V5-specific lives in {@link PerformanceProducerBase}; the v4
+ * client is driven by {@link PerformanceProducerV4} under the {@code produce-v4} name.
  */
 @Command(name = "produce", description = "Test pulsar producer performance.")
-@CustomLog
-public class PerformanceProducer extends PerformanceTopicListArguments{
-    private final LongAdder messagesSent = new LongAdder();
-    private final LongAdder messagesFailed = new LongAdder();
-    private final LongAdder bytesSent = new LongAdder();
-
-    private final LongAdder totalNumTxnOpenTxnFail = new LongAdder();
-    private final LongAdder totalNumTxnOpenTxnSuccess = new LongAdder();
-
-    private final LongAdder totalMessagesSent = new LongAdder();
-    private final LongAdder totalBytesSent = new LongAdder();
-
-    // Publish latencies are recorded in microseconds. A send slower than this means the benchmark is
-    // broken rather than slow, so values are clamped to keep HdrHistogram in range.
-    private static final long MAX_LATENCY_MICROS = TimeUnit.HOURS.toMicros(1);
-
-    private final Recorder recorder =
-            new Recorder(MAX_LATENCY_MICROS, LATENCY_HISTOGRAM_SIGNIFICANT_DIGITS);
-    private final Recorder cumulativeRecorder =
-            new Recorder(MAX_LATENCY_MICROS, LATENCY_HISTOGRAM_SIGNIFICANT_DIGITS);
-
-    private final LongAdder totalEndTxnOpSuccessNum = new LongAdder();
-    private final LongAdder totalEndTxnOpFailNum = new LongAdder();
-    private final LongAdder numTxnOpSuccess = new LongAdder();
-
-    private static IMessageFormatter messageFormatter = null;
-
-    @Option(names = { "-threads", "--num-test-threads" }, description = "Number of test threads",
-            converter = PositiveNumberParameterConvert.class
-    )
-    public int numTestThreads = 1;
-
-    @Option(names = { "-r", "--rate" }, description = "Publish rate msg/s across topics")
-    public int msgRate = 100;
-
-    @Option(names = { "-s", "--size" }, description = "Message size (bytes)")
-    public int msgSize = 1024;
-
-    @Option(names = { "-n", "--num-producers" }, description = "Number of producers (per topic)",
-            converter = PositiveNumberParameterConvert.class
-    )
-    public int numProducers = 1;
-
-    @Option(names = {"--separator"}, description = "Separator between the topic and topic number")
-    public String separator = "-";
-
-    @Option(names = {"--send-timeout"}, description = "Set the sendTimeout value default 0 to keep "
-            + "compatibility with previous version of pulsar-perf")
-    public int sendTimeout = 0;
-
-    @Option(names = { "-pn", "--producer-name" }, description = "Producer Name")
-    public String producerName = null;
-
-    @Option(names = { "-au", "--admin-url" }, description = "Pulsar Admin URL", descriptionKey = "webServiceUrl")
-    public String adminURL;
-
-    @Option(names = { "-ch",
-            "--chunking" }, description = "Should split the message and publish in chunks if message size is "
-            + "larger than allowed max size")
-    private boolean chunkingAllowed = false;
-
-    @Option(names = { "-o", "--max-outstanding" }, description = "Max number of outstanding messages")
-    public int maxOutstanding = DEFAULT_MAX_PENDING_MESSAGES;
-
-    @Option(names = { "-p", "--max-outstanding-across-partitions" }, description = "Max number of outstanding "
-            + "messages across partitions")
-    public int maxPendingMessagesAcrossPartitions = DEFAULT_MAX_PENDING_MESSAGES_ACROSS_PARTITIONS;
-
-    @Option(names = { "-np", "--partitions" }, description = "Create partitioned topics with the given number "
-            + "of partitions, set 0 to not try to create the topic")
-    public Integer partitions = null;
-
-    @Option(names = { "-m",
-            "--num-messages" }, description = "Number of messages to publish in total. If <= 0, it will keep "
-            + "publishing")
-    public long numMessages = 0;
+public class PerformanceProducer
+        extends PerformanceProducerBase<PulsarClient, AsyncProducer<byte[]>, Transaction> {
 
     @Option(names = { "-z", "--compression" }, description = "Compress messages payload")
     public CompressionType compression = CompressionType.NONE;
 
-    @Option(names = { "-f", "--payload-file" }, description = "Use payload from an UTF-8 encoded text file and "
-            + "a payload will be randomly selected when publishing messages")
-    public String payloadFilename = null;
-
-    @Option(names = { "-e", "--payload-delimiter" }, description = "The delimiter used to split lines when "
-            + "using payload from a file")
-    // here escaping \n since default value will be printed with the help text
-    public String payloadDelimiter = "\\n";
-
-    @Option(names = { "-b",
-            "--batch-time-window" }, description = "Batch messages in 'x' ms window (Default: 1ms)")
-    public double batchTimeMillis = 1.0;
-
-    @Option(names = { "-db",
-            "--disable-batching" }, description = "Disable batching if true")
-    public boolean disableBatching;
-
-    @Option(names = {
-            "-bm", "--batch-max-messages"
-    }, description = "Maximum number of messages per batch")
-    public int batchMaxMessages = DEFAULT_BATCHING_MAX_MESSAGES;
-
-    @Option(names = {
-            "-bb", "--batch-max-bytes"
-    }, description = "Maximum number of bytes per batch")
-    public int batchMaxBytes = 4 * 1024 * 1024;
-
-    @Option(names = { "-time",
-            "--test-duration" }, description = "Test duration in secs. If <= 0, it will keep publishing")
-    public long testTime = 0;
-
-    @Option(names = "--warmup-time", description = "Warm-up time in seconds (Default: 1 sec)")
-    public double warmupTimeSeconds = 1.0;
-
-    @Option(names = { "-k", "--encryption-key-name" }, description = "The public key name to encrypt payload")
-    public String encKeyName = null;
-
-    @Option(names = { "-v",
-            "--encryption-key-value-file" },
-            description = "The file which contains the public key to encrypt payload")
-    public String encKeyFile = null;
-
-    @Option(names = { "-d",
-            "--delay" }, description = "Mark messages with a given delay in seconds")
-    public long delay = 0;
-
-    @Option(names = { "-dr", "--delay-range"}, description = "Mark messages with a given delay by a random"
-            + " number of seconds. this value between the specified origin (inclusive) and the specified bound"
-            + " (exclusive). e.g. 1,300", converter = RangeConvert.class)
-    public Range<Long> delayRange = null;
-
-    @Option(names = { "-set",
-            "--set-event-time" }, description = "Set the eventTime on messages")
-    public boolean setEventTime = false;
-
-    @Option(names = { "-ef",
-            "--exit-on-failure" }, description = "Exit from the process on publish failure (default: disable)")
-    public boolean exitOnFailure = false;
-
-    @Option(names = {"-mk", "--message-key-generation-mode"}, description = "The generation mode of message key"
-            + ", valid options are: [autoIncrement, random]", descriptionKey = "messageKeyGenerationMode")
-    public String messageKeyGenerationMode = null;
-
     @Option(names = { "-am", "--access-mode" }, description = "Producer access mode")
     public ProducerAccessMode producerAccessMode = ProducerAccessMode.SHARED;
-
-    @Option(names = { "-fp", "--format-payload" },
-            description = "Format %%i as a message index in the stream from producer and/or %%t as the timestamp"
-                    + " nanoseconds.")
-    public boolean formatPayload = false;
-
-    @Option(names = {"-fc", "--format-class"}, description = "Custom Formatter class name")
-    public String formatterClass = "org.apache.pulsar.testclient.DefaultMessageFormatter";
-
-    @Option(names = {"-tto", "--txn-timeout"}, description = "Set the time value of transaction timeout,"
-            + " and the time unit is second. (After --txn-enable setting to true, --txn-timeout takes effect)")
-    public long transactionTimeout = 10;
-
-    @Option(names = {"-nmt", "--numMessage-perTransaction"},
-            description = "The number of messages sent by a transaction. "
-                    + "(After --txn-enable setting to true, -nmt takes effect)")
-    public int numMessagesPerTransaction = 50;
-
-    @Option(names = {"-txn", "--txn-enable"}, description = "Enable or disable the transaction")
-    public boolean isEnableTransaction = false;
-
-    @Option(names = {"-abort"}, description = "Abort the transaction. (After --txn-enable "
-            + "setting to true, -abort takes effect)")
-    public boolean isAbortTransaction = false;
-
-    @Option(names = { "--histogram-file" }, description = "HdrHistogram output file")
-    public String histogramFile = null;
-
-    @Override
-    public void run() throws Exception {
-
-        // Dump config variables
-        PerfClientUtils.printJVMInformation(log);
-        ObjectMapper m = new ObjectMapper();
-        ObjectWriter w = m.writerWithDefaultPrettyPrinter();
-        log.info().attr("config", w.writeValueAsString(this)).log("Starting Pulsar perf producer with config");
-
-        // Read payload data from file if needed
-        final byte[] payloadBytes = new byte[msgSize];
-        Random random = new Random(0);
-        List<byte[]> payloadByteList = new ArrayList<>();
-        if (this.payloadFilename != null) {
-            Path payloadFilePath = Paths.get(this.payloadFilename);
-            if (Files.notExists(payloadFilePath) || Files.size(payloadFilePath) == 0)  {
-                throw new IllegalArgumentException("Payload file doesn't exist or it is empty.");
-            }
-            // here escaping the default payload delimiter to correct value
-            String delimiter = this.payloadDelimiter.equals("\\n") ? "\n" : this.payloadDelimiter;
-            String[] payloadList = new String(Files.readAllBytes(payloadFilePath),
-                    StandardCharsets.UTF_8).split(delimiter);
-            log.info()
-                    .attr("payloads", payloadFilePath.toAbsolutePath())
-                    .attr("length", payloadList.length)
-                    .log("Reading payloads from and records read");
-            for (String payload : payloadList) {
-                payloadByteList.add(payload.getBytes(StandardCharsets.UTF_8));
-            }
-
-            if (this.formatPayload) {
-                messageFormatter = getMessageFormatter(this.formatterClass);
-            }
-        } else {
-            for (int i = 0; i < payloadBytes.length; ++i) {
-                payloadBytes[i] = (byte) (random.nextInt(26) + 65);
-            }
-        }
-
-        long start = System.nanoTime();
-
-        ExecutorService executor = Executors
-                .newCachedThreadPool(new DefaultThreadFactory("pulsar-perf-producer-exec"));
-        Thread shutdownHookThread = PerfClientUtils.addShutdownHook(() -> {
-            executorShutdownNow(executor);
-            printAggregatedThroughput(start);
-            printAggregatedStats();
-        });
-
-        if (this.partitions  != null) {
-            final PulsarAdminBuilder adminBuilder = PerfClientUtils
-                    .createAdminBuilderFromArguments(this, this.adminURL);
-
-            try (PulsarAdmin adminClient = adminBuilder.build()) {
-                for (String topic : this.topics) {
-                    log.info()
-                            .attr("topic", topic)
-                            .attr("partitions", this.partitions)
-                            .log("Creating partitioned topic with partitions");
-                    try {
-                        adminClient.topics().createPartitionedTopic(topic, this.partitions);
-                    } catch (PulsarAdminException.ConflictException alreadyExists) {
-                        log.debug().attr("topic", topic).attr("exists", alreadyExists).log("Topic already exists");
-                        PartitionedTopicMetadata partitionedTopicMetadata = adminClient.topics()
-                                .getPartitionedTopicMetadata(topic);
-                        if (partitionedTopicMetadata.partitions != this.partitions) {
-                            log.error()
-                                    .attr("topic", topic)
-                                    .attr("partitions", partitionedTopicMetadata.partitions)
-                                    .attr("expecting", this.partitions)
-                                    .log("Topic  already exists but it has a wrong number of partitions: , expecting");
-                            PerfClientUtils.exit(1);
-                        }
-                    }
-                }
-            }
-        }
-
-        CountDownLatch doneLatch = new CountDownLatch(this.numTestThreads);
-
-        final long numMessagesPerThread = this.numMessages / this.numTestThreads;
-        final int msgRatePerThread = this.msgRate / this.numTestThreads;
-
-        for (int i = 0; i < this.numTestThreads; i++) {
-            final int threadIdx = i;
-            executor.submit(() -> {
-                log.info().attr("thread", threadIdx).log("Started performance test thread");
-                runProducer(
-                        threadIdx,
-                        this,
-                        numMessagesPerThread,
-                        msgRatePerThread,
-                        payloadByteList,
-                        payloadBytes,
-                        doneLatch
-                );
-            });
-        }
-
-        // Print report stats
-        long oldTime = System.nanoTime();
-
-        Histogram reportHistogram = null;
-        HistogramLogWriter histogramLogWriter = null;
-
-        if (this.histogramFile != null) {
-            String statsFileName = this.histogramFile;
-            log.info().attr("stats", statsFileName).log("Dumping latency stats to");
-
-            PrintStream histogramLog = new PrintStream(new FileOutputStream(statsFileName), false);
-            histogramLogWriter = new HistogramLogWriter(histogramLog);
-
-            // Some log header bits
-            histogramLogWriter.outputLogFormatVersion();
-            histogramLogWriter.outputLegend();
-        }
-
-        while (!Thread.currentThread().isInterrupted()) {
-            try {
-                Thread.sleep(10000);
-            } catch (InterruptedException e) {
-                Thread.currentThread().interrupt();
-                break;
-            }
-
-            if (doneLatch.getCount() <= 0) {
-                break;
-            }
-
-            long now = System.nanoTime();
-            double elapsed = (now - oldTime) / 1e9;
-            long total = totalMessagesSent.sum();
-            long totalTxnOpSuccess = 0;
-            long totalTxnOpFail = 0;
-            double rateOpenTxn = 0;
-            double rate = messagesSent.sumThenReset() / elapsed;
-            double failureRate = messagesFailed.sumThenReset() / elapsed;
-            double throughput = bytesSent.sumThenReset() / elapsed / 1024 / 1024 * 8;
-
-            reportHistogram = recorder.getIntervalHistogram(reportHistogram);
-
-            if (this.isEnableTransaction) {
-                totalTxnOpSuccess = totalEndTxnOpSuccessNum.sum();
-                totalTxnOpFail = totalEndTxnOpFailNum.sum();
-                rateOpenTxn = numTxnOpSuccess.sumThenReset() / elapsed;
-                log.infof("--- Transaction: %d transaction end successfully"
-                                + " --- %d transaction end failed --- %.3f Txn/s",
-                        totalTxnOpSuccess, totalTxnOpFail, rateOpenTxn);
-            }
-            log.infof("Throughput produced: %7d msg --- %8.1f msg/s --- %8.1f Mbit/s"
-                            + " --- failure %8.1f msg/s"
-                            + " --- Latency: mean: %7.3f ms - med: %7.3f"
-                            + " - 95pct: %7.3f - 99pct: %7.3f"
-                            + " - 99.9pct: %7.3f - 99.99pct: %7.3f - Max: %7.3f",
-                    total, rate, throughput, failureRate,
-                    reportHistogram.getMean() / 1000.0,
-                    reportHistogram.getValueAtPercentile(50) / 1000.0,
-                    reportHistogram.getValueAtPercentile(95) / 1000.0,
-                    reportHistogram.getValueAtPercentile(99) / 1000.0,
-                    reportHistogram.getValueAtPercentile(99.9) / 1000.0,
-                    reportHistogram.getValueAtPercentile(99.99) / 1000.0,
-                    reportHistogram.getMaxValue() / 1000.0);
-
-            if (histogramLogWriter != null) {
-                histogramLogWriter.outputIntervalHistogram(reportHistogram);
-            }
-
-            reportHistogram.reset();
-
-            oldTime = now;
-        }
-
-        PerfClientUtils.removeAndRunShutdownHook(shutdownHookThread);
-    }
 
     public PerformanceProducer() {
         super("produce");
     }
 
-    private static void executorShutdownNow(ExecutorService executor) {
-        executor.shutdownNow();
-        try {
-            if (!executor.awaitTermination(10, TimeUnit.SECONDS)) {
-                log.warn("Failed to terminate executor within timeout. The following are stack"
-                        + " traces of still running threads.");
-            }
-        } catch (InterruptedException e) {
-            log.warn("Shutdown of thread pool was interrupted");
-            Thread.currentThread().interrupt();
+    @Override
+    protected PulsarClient createClient() throws PulsarClientException {
+        PulsarClientBuilder clientBuilder = PerfClientUtils.createV5ClientBuilderFromArguments(this);
+        if (this.isEnableTransaction) {
+            clientBuilder.transactionPolicy(TransactionPolicy.builder()
+                    .timeout(Duration.ofSeconds(this.transactionTimeout))
+                    .build());
         }
+        return clientBuilder.build();
     }
 
-    @SuppressWarnings("unchecked")
-    static IMessageFormatter getMessageFormatter(String formatterClass) {
-        try {
-            ClassLoader classLoader = PerformanceProducer.class.getClassLoader();
-            Class clz = classLoader.loadClass(formatterClass);
-            return (IMessageFormatter) clz.getDeclaredConstructor().newInstance();
-        } catch (Exception e) {
-            if (PerfClientUtils.hasInterruptedException(e)) {
-                Thread.currentThread().interrupt();
-            }
-            return null;
-        }
+    @Override
+    protected void closeClient(PulsarClient client) {
+        PerfClientUtils.closeClient(client);
     }
 
     ProducerBuilder<byte[]> createProducerBuilder(PulsarClient client, int producerId, String topic) {
@@ -479,7 +93,8 @@ public class PerformanceProducer extends PerformanceTopicListArguments{
         // messageRoutingMode as user-configurable knobs; the SDK manages memory via the
         // client-level MemorySize policy and routes appropriately for regular and scalable
         // topics. The legacy --max-outstanding / --max-outstanding-across-partitions flags
-        // are accepted for back-compat but have no effect on the V5 client.
+        // are accepted for back-compat but have no effect on the V5 client. Use produce-v4
+        // to exercise them.
 
         if (this.producerName != null) {
             producerBuilder.producerName(String.format("%s%s%d", this.producerName, this.separator, producerId));
@@ -506,7 +121,7 @@ public class PerformanceProducer extends PerformanceTopicListArguments{
 
         if (isNotBlank(this.encKeyName) && isNotBlank(this.encKeyFile)) {
             PemFileKeyProvider keyProvider = PemFileKeyProvider.builder()
-                    .publicKey(this.encKeyName, java.nio.file.Path.of(this.encKeyFile))
+                    .publicKey(this.encKeyName, Path.of(this.encKeyFile))
                     .build();
             producerBuilder.encryptionPolicy(ProducerEncryptionPolicy.builder()
                     .publicKeyProvider(keyProvider)
@@ -517,333 +132,54 @@ public class PerformanceProducer extends PerformanceTopicListArguments{
         return producerBuilder;
     }
 
-    private void runProducer(int producerId,
-                                    PerformanceProducer arguments,
-                                    long numMessages,
-                                    int msgRate,
-                                    List<byte[]> payloadByteList,
-                                    byte[] payloadBytes,
-                                    CountDownLatch doneLatch) {
-        PulsarClient client = null;
-        boolean produceEnough = false;
-        try {
-            List<Future<Producer<byte[]>>> futures = new ArrayList<>();
+    @Override
+    protected CompletableFuture<AsyncProducer<byte[]>> createProducerAsync(PulsarClient client, int producerId,
+                                                                           String topic) {
+        return createProducerBuilder(client, producerId, topic).createAsync().thenApply(Producer::async);
+    }
 
-            PulsarClientBuilder clientBuilder = PerfClientUtils.createV5ClientBuilderFromArguments(arguments);
-            if (this.isEnableTransaction) {
-                clientBuilder.transactionPolicy(TransactionPolicy.builder()
-                        .timeout(Duration.ofSeconds(this.transactionTimeout))
-                        .build());
-            }
-            client = clientBuilder.build();
+    @Override
+    protected Transaction newTransaction(PulsarClient client) throws PulsarClientException {
+        return client.newTransaction();
+    }
 
-            AtomicReference<Transaction> transactionAtomicReference;
-            if (this.isEnableTransaction) {
-                transactionAtomicReference = new AtomicReference<>(
-                        PerfClientUtils.newTransactionWithRetry(client));
-            } else {
-                transactionAtomicReference = new AtomicReference<>(null);
-            }
+    @Override
+    protected Transaction openFirstTransaction(PulsarClient client)
+            throws PulsarClientException, InterruptedException {
+        return PerfClientUtils.newTransactionWithRetry(client);
+    }
 
-            for (int i = 0; i < this.numTopics; i++) {
+    @Override
+    protected CompletableFuture<Void> commitTransaction(Transaction transaction) {
+        return transaction.async().commit();
+    }
 
-                String topic = this.topics.get(i);
-                log.info().attr("adding", this.numProducers).attr("topic", topic).log("Adding publishers on topic");
+    @Override
+    protected CompletableFuture<Void> abortTransaction(Transaction transaction) {
+        return transaction.async().abort();
+    }
 
-                for (int j = 0; j < this.numProducers; j++) {
-                    ProducerBuilder<byte[]> prodBuilder = createProducerBuilder(client, producerId, topic);
-                    futures.add(prodBuilder.createAsync());
-                }
-            }
-
-            final List<Producer<byte[]>> producers = new ArrayList<>(futures.size());
-            for (Future<Producer<byte[]>> future : futures) {
-                producers.add(future.get());
-            }
-            Collections.shuffle(producers);
-            final List<AsyncProducer<byte[]>> asyncProducers = new ArrayList<>(producers.size());
-            for (Producer<byte[]> p : producers) {
-                asyncProducers.add(p.async());
-            }
-
-            log.info().attr("created", producers.size()).log("Created producers");
-
-            RateLimiter rateLimiter = RateLimiter.create(msgRate);
-
-            long startTime = System.nanoTime();
-            long warmupEndTime = startTime + (long) (this.warmupTimeSeconds * 1e9);
-            long testEndTime = startTime + (long) (this.testTime * 1e9);
-            MessageKeyGenerationMode msgKeyMode = null;
-            if (isNotBlank(this.messageKeyGenerationMode)) {
-                try {
-                    msgKeyMode = MessageKeyGenerationMode.valueOf(this.messageKeyGenerationMode);
-                } catch (IllegalArgumentException e) {
-                    throw new IllegalArgumentException("messageKeyGenerationMode only support [autoIncrement, random]");
-                }
-            }
-            // Send messages on all topics/producers
-            AtomicLong totalSent = new AtomicLong(0);
-            AtomicLong numMessageSend = new AtomicLong(0);
-            Semaphore numMsgPerTxnLimit = new Semaphore(this.numMessagesPerTransaction);
-            // Send futures of the in-flight transaction. V5 transaction-aware sends are queued
-            // onto an internal dispatch chain, so the v4-side txn-coordinator registration can
-            // lag the local counter; we await these before committing so commit never races
-            // ahead of the sends (otherwise the broker rejects with InvalidTxnStatusException).
-            final List<java.util.concurrent.CompletableFuture<?>> pendingTxnSends = new ArrayList<>();
-            while (!Thread.currentThread().isInterrupted()) {
-                if (produceEnough) {
-                    break;
-                }
-                for (AsyncProducer<byte[]> producer : asyncProducers) {
-                    if (this.testTime > 0) {
-                        if (System.nanoTime() > testEndTime) {
-                            log.info()
-                                    .attr("duration", this.testTime)
-                                    .log("------------- DONE (reached the maximum duration:"
-                                            + " [ seconds] of production) --------------");
-                            doneLatch.countDown();
-                            produceEnough = true;
-                            break;
-                        }
-                    }
-
-                    if (numMessages > 0) {
-                        if (totalSent.get() >= numMessages) {
-                            log.info()
-                                    .attr("number", numMessages)
-                                    .log("DONE (reached the maximum number: of production");
-                            doneLatch.countDown();
-                            produceEnough = true;
-                            break;
-                        }
-                    }
-                    rateLimiter.acquire();
-                    //if transaction is disable, transaction will be null.
-                    Transaction transaction = transactionAtomicReference.get();
-                    final long sendTime = System.nanoTime();
-
-                    byte[] payloadData;
-
-                    if (this.payloadFilename != null) {
-                        if (messageFormatter != null) {
-                            payloadData = messageFormatter.formatMessage(this.producerName, totalSent.get(),
-                                    payloadByteList.get(ThreadLocalRandom.current().nextInt(payloadByteList.size())));
-                        } else {
-                            payloadData = payloadByteList.get(
-                                    ThreadLocalRandom.current().nextInt(payloadByteList.size()));
-                        }
-                    } else {
-                        payloadData = payloadBytes;
-                    }
-                    AsyncMessageBuilder<byte[]> messageBuilder = producer.newMessage().value(payloadData);
-                    if (this.isEnableTransaction) {
-                        if (this.numMessagesPerTransaction > 0) {
-                            try {
-                                numMsgPerTxnLimit.acquire();
-                            } catch (InterruptedException exception){
-                                log.error().exception(exception).log("Get exception");
-                                Thread.currentThread().interrupt();
-                            }
-                        }
-                        messageBuilder.transaction(transaction);
-                    }
-                    if (this.delay > 0) {
-                        messageBuilder.deliverAfter(Duration.ofSeconds(this.delay));
-                    } else if (this.delayRange != null) {
-                        final long deliverAfter = ThreadLocalRandom.current()
-                                .nextLong(this.delayRange.lowerEndpoint(), this.delayRange.upperEndpoint());
-                        messageBuilder.deliverAfter(Duration.ofSeconds(deliverAfter));
-                    }
-                    if (this.setEventTime) {
-                        messageBuilder.eventTime(Instant.now());
-                    }
-                    //generate msg key
-                    if (msgKeyMode == MessageKeyGenerationMode.random) {
-                        messageBuilder.key(String.valueOf(ThreadLocalRandom.current().nextInt()));
-                    } else if (msgKeyMode == MessageKeyGenerationMode.autoIncrement) {
-                        messageBuilder.key(String.valueOf(totalSent.get()));
-                    }
-                    PulsarClient pulsarClient = client;
-                    var sendFuture = messageBuilder.send().thenRun(() -> {
-                        bytesSent.add(payloadData.length);
-                        messagesSent.increment();
-                        totalSent.incrementAndGet();
-                        totalMessagesSent.increment();
-                        totalBytesSent.add(payloadData.length);
-
-                        long now = System.nanoTime();
-                        if (now > warmupEndTime) {
-                            long latencyMicros =
-                                    Math.min(NANOSECONDS.toMicros(now - sendTime), MAX_LATENCY_MICROS);
-                            recorder.recordValue(latencyMicros);
-                            cumulativeRecorder.recordValue(latencyMicros);
-                        }
-                    }).exceptionally(ex -> {
-                        Throwable cause = FutureUtil.unwrapCompletionException(ex);
-                        // Ignore the exception when the producer is closed
-                        if (cause instanceof PulsarClientException.AlreadyClosedException) {
-                            return null;
-                        }
-                        if (PerfClientUtils.hasInterruptedException(ex)) {
-                            Thread.currentThread().interrupt();
-                            return null;
-                        }
-                        log.warn().exception(ex).log("Write message error with exception");
-                        messagesFailed.increment();
-                        if (this.exitOnFailure) {
-                            PerfClientUtils.exit(1);
-                        }
-                        return null;
-                    });
-                    if (this.isEnableTransaction) {
-                        pendingTxnSends.add(sendFuture);
-                    }
-                    if (this.isEnableTransaction
-                            && numMessageSend.incrementAndGet() == this.numMessagesPerTransaction) {
-                        // Await all sends issued under this transaction before committing, so the
-                        // txn coordinator has registered every send. The chain above already
-                        // swallows per-send failures, so this join never throws on a send error.
-                        try {
-                            java.util.concurrent.CompletableFuture.allOf(
-                                    pendingTxnSends.toArray(new java.util.concurrent.CompletableFuture[0]))
-                                    .join();
-                        } catch (Exception awaitEx) {
-                            if (PerfClientUtils.hasInterruptedException(awaitEx)) {
-                                Thread.currentThread().interrupt();
-                            }
-                        }
-                        pendingTxnSends.clear();
-                        if (!this.isAbortTransaction) {
-                            transaction.async().commit()
-                                    .thenRun(() -> {
-                                        log.debug().log("Committed transaction");
-                                        totalEndTxnOpSuccessNum.increment();
-                                        numTxnOpSuccess.increment();
-                                    })
-                                    .exceptionally(exception -> {
-                                        if (PerfClientUtils.hasInterruptedException(exception)) {
-                                            Thread.currentThread().interrupt();
-                                            return null;
-                                        }
-                                        log.error()
-                                                .exception(exception)
-                                                .log("Commit transaction failed with exception");
-                                        totalEndTxnOpFailNum.increment();
-                                        return null;
-                                    });
-                        } else {
-                            transaction.async().abort().thenRun(() -> {
-                                log.debug().log("Abort transaction");
-                                totalEndTxnOpSuccessNum.increment();
-                                numTxnOpSuccess.increment();
-                            }).exceptionally(exception -> {
-                                if (PerfClientUtils.hasInterruptedException(exception)) {
-                                    Thread.currentThread().interrupt();
-                                    return null;
-                                }
-                                log.error()
-                                        .exception(exception)
-                                        .log("Abort transaction failed with exception");
-                                totalEndTxnOpFailNum.increment();
-                                return null;
-                            });
-                        }
-                        while (!Thread.currentThread().isInterrupted()) {
-                            try {
-                                Transaction newTransaction = pulsarClient.newTransaction();
-                                transactionAtomicReference.compareAndSet(transaction, newTransaction);
-                                numMessageSend.set(0);
-                                numMsgPerTxnLimit.release(this.numMessagesPerTransaction);
-                                totalNumTxnOpenTxnSuccess.increment();
-                                break;
-                            } catch (Exception e){
-                                if (PerfClientUtils.hasInterruptedException(e)) {
-                                    Thread.currentThread().interrupt();
-                                } else {
-                                    totalNumTxnOpenTxnFail.increment();
-                                    log.error().exception(e).log("Failed to new transaction with exception");
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        } catch (Throwable t) {
-            if (PerfClientUtils.hasInterruptedException(t)) {
-                Thread.currentThread().interrupt();
-            } else {
-                log.error().exception(t).log("Got error");
-            }
-        } finally {
-            if (!produceEnough) {
-                doneLatch.countDown();
-            }
-            PerfClientUtils.closeClient(client);
+    @Override
+    protected CompletableFuture<?> sendMessage(AsyncProducer<byte[]> producer, byte[] payload,
+                                               Transaction transaction, String key, Long deliverAfterSeconds) {
+        AsyncMessageBuilder<byte[]> messageBuilder = producer.newMessage().value(payload);
+        if (transaction != null) {
+            messageBuilder.transaction(transaction);
         }
-    }
-
-    private void printAggregatedThroughput(long start) {
-        double elapsed = (System.nanoTime() - start) / 1e9;
-        double rate = totalMessagesSent.sum() / elapsed;
-        double throughput = totalBytesSent.sum() / elapsed / 1024 / 1024 * 8;
-        long totalTxnSuccess = 0;
-        long totalTxnFail = 0;
-        double rateOpenTxn = 0;
-        long numTransactionOpenFailed = 0;
-        long numTransactionOpenSuccess = 0;
-
-        if (this.isEnableTransaction) {
-            totalTxnSuccess = totalEndTxnOpSuccessNum.sum();
-            totalTxnFail = totalEndTxnOpFailNum.sum();
-            rateOpenTxn = elapsed / (totalTxnFail + totalTxnSuccess);
-            numTransactionOpenFailed = totalNumTxnOpenTxnFail.sum();
-            numTransactionOpenSuccess = totalNumTxnOpenTxnSuccess.sum();
-            log.infof("--- Transaction: %d transaction end successfully"
-                            + " --- %d transaction end failed"
-                            + " --- %d transaction open successfully"
-                            + " --- %d transaction open failed --- %.3f Txn/s",
-                    totalTxnSuccess, totalTxnFail,
-                    numTransactionOpenSuccess, numTransactionOpenFailed, rateOpenTxn);
+        if (deliverAfterSeconds != null) {
+            messageBuilder.deliverAfter(Duration.ofSeconds(deliverAfterSeconds));
         }
-        log.infof("Aggregated throughput stats --- %d records sent --- %.3f msg/s --- %.3f Mbit/s",
-                totalMessagesSent.sum(), rate, throughput);
-    }
-
-    private void printAggregatedStats() {
-        Histogram reportHistogram = cumulativeRecorder.getIntervalHistogram();
-
-        log.infof("Aggregated latency stats --- Latency: mean: %7.3f ms"
-                        + " - med: %7.3f - 95pct: %7.3f - 99pct: %7.3f"
-                        + " - 99.9pct: %7.3f - 99.99pct: %7.3f"
-                        + " - 99.999pct: %7.3f - Max: %7.3f",
-                reportHistogram.getMean() / 1000.0,
-                reportHistogram.getValueAtPercentile(50) / 1000.0,
-                reportHistogram.getValueAtPercentile(95) / 1000.0,
-                reportHistogram.getValueAtPercentile(99) / 1000.0,
-                reportHistogram.getValueAtPercentile(99.9) / 1000.0,
-                reportHistogram.getValueAtPercentile(99.99) / 1000.0,
-                reportHistogram.getValueAtPercentile(99.999) / 1000.0,
-                reportHistogram.getMaxValue() / 1000.0);
-    }
-
-    public enum MessageKeyGenerationMode {
-        autoIncrement, random
-    }
-
-    static class RangeConvert implements ITypeConverter<Range<Long>> {
-        @Override
-        public Range<Long> convert(String rangeStr) {
-            try {
-                requireNonNull(rangeStr);
-                final String[] facts = rangeStr.split(",");
-                final long min = Long.parseLong(facts[0].trim());
-                final long max = Long.parseLong(facts[1].trim());
-                return Range.closedOpen(min, max);
-            } catch (Throwable ex) {
-                throw new TypeConversionException("Unknown delay range interval,"
-                        + " the format should be \"<origin>,<bound>\". error message: " + rangeStr);
-            }
+        if (this.setEventTime) {
+            messageBuilder.eventTime(Instant.now());
         }
+        if (key != null) {
+            messageBuilder.key(key);
+        }
+        return messageBuilder.send();
     }
 
+    @Override
+    protected boolean isAlreadyClosedException(Throwable cause) {
+        return cause instanceof PulsarClientException.AlreadyClosedException;
+    }
 }
