+++ tests/integration/src/test/java/org/apache/pulsar/tests/integration/cli/PerfToolTest.java
@@ -70,9 +70,50 @@ public class PerfToolTest extends TopicMessagingBase {
                 "PerformanceReader - Aggregated latency stats");
     }
 
+    @Test
+    public void testProduceV4() throws Exception {
+        String serviceUrl = "pulsar://" + pulsarCluster.getProxy().getContainerName() + ":"
+                + PulsarContainer.BROKER_PORT;
+        final String topicName = getNonPartitionedTopic("testProduceV4", true);
+        ZKContainer clientToolContainer = pulsarCluster.getZooKeeper();
+        ContainerExecResult produceResult =
+                produceWithPerfTool(clientToolContainer, "produce-v4", serviceUrl, topicName, MESSAGE_COUNT);
+        checkOutputForLogs(produceResult, "PerformanceProducerV4 - Aggregated throughput stats",
+                "PerformanceProducerV4 - Aggregated latency stats");
+    }
+
+    @Test
+    public void testConsumeV4() throws Exception {
+        String serviceUrl = "pulsar://" + pulsarCluster.getProxy().getContainerName() + ":"
+                + PulsarContainer.BROKER_PORT;
+        final String topicName = getNonPartitionedTopic("testConsumeV4", true);
+        ZKContainer clientToolContainer = pulsarCluster.getZooKeeper();
+        ContainerExecResult consumeResult =
+                consumeWithPerfTool(clientToolContainer, "consume-v4", "produce-v4", serviceUrl, topicName);
+        checkOutputForLogs(consumeResult, "PerformanceConsumerV4 - Aggregated throughput stats",
+                "PerformanceConsumerV4 - Aggregated latency stats");
+    }
+
+    @Test
+    public void testReadV4() throws Exception {
+        String serviceUrl = "pulsar://" + pulsarCluster.getProxy().getContainerName() + ":"
+                + PulsarContainer.BROKER_PORT;
+        final String topicName = getNonPartitionedTopic("testReadV4", true);
+        ZKContainer clientToolContainer = pulsarCluster.getZooKeeper();
+        ContainerExecResult readResult =
+                readWithPerfTool(clientToolContainer, "read-v4", "produce-v4", serviceUrl, topicName);
+        checkOutputForLogs(readResult, "PerformanceReaderV4 - Aggregated throughput stats ",
+                "PerformanceReaderV4 - Aggregated latency stats");
+    }
+
     private ContainerExecResult produceWithPerfTool(ChaosContainer<?> container, String url, String topic,
                                                     int messageCount) throws Exception {
-        ContainerExecResult result = container.execCmd("bin/pulsar-perf", "produce", "-u", url, "-m",
+        return produceWithPerfTool(container, "produce", url, topic, messageCount);
+    }
+
+    private ContainerExecResult produceWithPerfTool(ChaosContainer<?> container, String command, String url,
+                                                    String topic, int messageCount) throws Exception {
+        ContainerExecResult result = container.execCmd("bin/pulsar-perf", command, "-u", url, "-m",
                 String.valueOf(messageCount), topic);
 
         return failOnError("Performance producer", result);
@@ -80,9 +121,15 @@ public class PerfToolTest extends TopicMessagingBase {
 
     private ContainerExecResult consumeWithPerfTool(ChaosContainer<?> container, String url, String topic)
             throws Exception {
-        CompletableFuture<ContainerExecResult> resultFuture = container.execCmdAsync("bin/pulsar-perf", "consume", "-u",
-                url, "-m", String.valueOf(MESSAGE_COUNT), topic);
-        produceWithPerfTool(container, url, topic, MESSAGE_COUNT);
+        return consumeWithPerfTool(container, "consume", "produce", url, topic);
+    }
+
+    private ContainerExecResult consumeWithPerfTool(ChaosContainer<?> container, String consumeCommand,
+                                                    String produceCommand, String url, String topic)
+            throws Exception {
+        CompletableFuture<ContainerExecResult> resultFuture = container.execCmdAsync("bin/pulsar-perf",
+                consumeCommand, "-u", url, "-m", String.valueOf(MESSAGE_COUNT), topic);
+        produceWithPerfTool(container, produceCommand, url, topic, MESSAGE_COUNT);
 
         ContainerExecResult result = resultFuture.get(5, TimeUnit.SECONDS);
         return failOnError("Performance consumer", result);
@@ -90,9 +137,15 @@ public class PerfToolTest extends TopicMessagingBase {
 
     private ContainerExecResult readWithPerfTool(ChaosContainer<?> container, String url, String topic)
             throws Exception {
-        CompletableFuture<ContainerExecResult> resultFuture = container.execCmdAsync("bin/pulsar-perf", "read", "-u",
-                url, "-n", String.valueOf(MESSAGE_COUNT), topic);
-        produceWithPerfTool(container, url, topic, MESSAGE_COUNT);
+        return readWithPerfTool(container, "read", "produce", url, topic);
+    }
+
+    private ContainerExecResult readWithPerfTool(ChaosContainer<?> container, String readCommand,
+                                                 String produceCommand, String url, String topic)
+            throws Exception {
+        CompletableFuture<ContainerExecResult> resultFuture = container.execCmdAsync("bin/pulsar-perf",
+                readCommand, "-u", url, "-n", String.valueOf(MESSAGE_COUNT), topic);
+        produceWithPerfTool(container, produceCommand, url, topic, MESSAGE_COUNT);
 
         ContainerExecResult result = resultFuture.get(5, TimeUnit.SECONDS);
         return failOnError("Performance consumer", result);
